move some helpers to binary
