+++ src/util/binary.ts
@@ -1,0 +1,33 @@
+export const createHash = async (encodedData: string | Uint8Array<ArrayBufferLike>, algorithm = 'SHA-256') => {
+    const encoder = new TextEncoder();
+    encodedData = typeof encodedData === 'string' ? encoder.encode(encodedData) : encodedData;
+    const hashBuffer = await crypto.subtle.digest(algorithm, encodedData);
+    const hashArray = Array.from(new Uint8Array(hashBuffer));
+    return hashArray.map(byte => byte.toString(16).padStart(2, '0')).join('');
+};
+
+export const fileToBytes = (file: File): Promise<Uint8Array> => {
+    return new Promise((resolve, reject) => {
+        const reader = new FileReader();
+        reader.onload = () => {
+            resolve(new Uint8Array(reader.result as ArrayBuffer));
+        };
+        reader.onerror = reject;
+        reader.readAsArrayBuffer(file);
+    });
+};
+
+export const blobToBase64 = (blob: Blob): Promise<string> => {
+    return new Promise((resolve, reject) => {
+        const reader = new FileReader();
+        reader.onloadend = () => {
+            if (typeof reader.result === 'string') {
+                resolve(reader.result);
+            } else {
+                reject('Failed to convert blob to base64');
+            }
+        };
+        reader.onerror = reject;
+        reader.readAsDataURL(blob);
+    });
+};
